OPE: organize modules to avoid type errors
